requirements.txt
Django==4.1.7
environs==9.5.0
djangorestframework==3.14.0
django-rest-knox==4.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
